requirements.txt
# Python dependencies for Car Lease/Loan AI Assistant - OCR Module

# OCR Dependencies (Tesseract)
pytesseract>=0.3.10
PyMuPDF>=1.24.0  
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# PDF Text Extraction (native PDFs)
pdfplumber>=0.10.0

# Milestone 2: Image Preprocessing (Computer Vision)
opencv-python>=4.8.0

# Milestone 2: Database Integration
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Milestone 3+4: Unified Web Server
flask>=3.0.0
flask-cors>=4.0.0

# Milestone 3: AI Chatbot (Google Gemini)
google-generativeai>=0.3.0
